Split Routes into container and component files
